Add files via upload
We will be using the following two learning algorithms: 

1. Artificial neural networks (ANN) 
To start, we download any neural networks package to classify our classification problems. Then we experiment with the number of layers and number of nodes, activation functions (sigmoid, tanh, etc.), and maybe a couple of other parameters. 

2. K Nearest Neighbors (KNN)
Then we download and use any KNN package to classify our classification problems. We experiment with the number of neighbors.